requirements.txt
Pillow==10.2.0
colorthief==0.2.1
webcolors==1.13 